% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_CEPAL.R
\name{series_CEPAL}
\alias{series_CEPAL}
\title{Uma API para os dados abertos da CEPAL}
\usage{
series_CEPAL()
}
\description{
Esta função retorna uma tabela com o nome e id de todas as series disponíveis nos dados abertos da CEPAL
}
\examples{
series <- series_CEPAL()
}
\keyword{CEPAL}
\keyword{dados}
