% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPAL_stat.R
\name{CEPAL_stat}
\alias{CEPAL_stat}
\title{Uma API para os dados abertos da CEPAL}
\usage{
CEPAL_stat(id_serie)
}
\arguments{
\item{id_serie}{ID da série que será baixada. Para conhecer as séries disponíveis e seus IDs, use a função XXXX.}
}
\description{
Esta função faz download das séries da CEPAL
}
\examples{
termos_troca <- CEPAL_stat(883)
}
\keyword{CEPAL}
\keyword{dados}
